shared prefs repo fix const commit
