+++ app/src/main/java/my/doctrina/app/data/repository/SharedPreferencesRepository.kt
@@ -2,12 +2,16 @@ package my.doctrina.app.data.repository
 
 import android.content.Context
 import android.content.SharedPreferences
+import my.doctrina.app.data.api.response.SignInConfirmResponse
+
+private const val EMAIL_KEY = "email"
 
 class SharedPreferencesRepository(context: Context) {
 
     private val sharedPreferences: SharedPreferences =
         context.getSharedPreferences("user_prefs", Context.MODE_PRIVATE)
 
+    // TODO: imageUrl and email change to get|set sting and to do constants with key values
     fun saveImageUrl(imageUrlKey: String, urlValue: String) {
         sharedPreferences.edit().putString(imageUrlKey, urlValue).apply()
     }
@@ -16,21 +20,23 @@ class SharedPreferencesRepository(context: Context) {
         return sharedPreferences.getString(imageUrlKey, defaultUrlValue) ?: defaultUrlValue
     }
 
-    fun saveEmail(emailKey: String, emailValue: String) {
-        sharedPreferences.edit().putString(emailKey, emailValue).apply()
+    fun saveEmail(emailValue: String) {
+        sharedPreferences.edit().putString(EMAIL_KEY, emailValue).apply()
     }
 
-    fun getEmail(emailKey: String, defaultEmailValue: String): String {
-        return sharedPreferences.getString(emailKey, defaultEmailValue) ?: defaultEmailValue
+    fun getEmail(): String {
+        return sharedPreferences.getString(EMAIL_KEY, "") ?: ""
     }
 
-    fun saveUserData(
+    private fun saveUserData(
         accessExpired: Int,
         accessToken: String,
         refreshExpired: Int,
         refreshToken: String,
         success: Boolean?
     ) {
+
+        // TODO: constants with more than one keys
         val editor = sharedPreferences.edit()
         editor.putInt("access_expired", accessExpired)
         editor.putString("access_token", accessToken)
@@ -42,10 +48,33 @@ class SharedPreferencesRepository(context: Context) {
         editor.apply()
     }
 
+    fun getUserData(body: SignInConfirmResponse?) {
+        if (body != null) {
+            with(body) {
+                val accessExpired = accessExpired
+                val accessToken = accessToken.toString()
+                val refreshExpired = refreshExpired
+                val refreshToken = refreshToken.toString()
+                val success = success
+
+                if (accessExpired != null && refreshExpired != null) {
+                    saveUserData(
+                        accessExpired,
+                        accessToken,
+                        refreshExpired,
+                        refreshToken,
+                        success
+                    )
+                }
+            }
+        }
+    }
+
     fun getAccessExpirationUserData(): Int {
         return sharedPreferences.getInt("access_expired", 0)
     }
 
+    // TODO: save method gets token value. fix it
     fun saveAccessUserToken() {
         sharedPreferences.edit().putString("access_token", "").apply()
     }
