+++ app/src/main/java/my/doctrina/app/presentation/fragments/LoginCodeConfirmationFragment.kt
@@ -42,7 +42,7 @@ class LoginCodeConfirmationFragment : Fragment(R.layout.fragment_login_code_conf
 
         sharedPreferencesRepository = SharedPreferencesRepository(requireContext())
 
-        emailSP = sharedPreferencesRepository.getEmail("email", "")
+        emailSP = sharedPreferencesRepository.getEmail()
         imageUrlSP = sharedPreferencesRepository.getImageUrl("imageUrl", "")
         codeNumberEditTextList = ArrayList()
 
@@ -162,21 +162,7 @@ class LoginCodeConfirmationFragment : Fragment(R.layout.fragment_login_code_conf
                     response: Response<SignInConfirmResponse>
                 ) {
                     if (response.isSuccessful) {
-
-                        getUserData(response.body())
-
-//                        val accessExpired = response.body()?.accessExpired
-//                        val refreshExpired = response.body()?.refreshExpired
-//                        if (accessExpired != null && refreshExpired != null) {
-//
-//                            sharedPreferencesRepository.saveUserData(
-//                                accessExpired,
-//                                response.body()?.accessToken.toString(),
-//                                refreshExpired,
-//                                response.body()?.refreshToken.toString(),
-//                                response.body()!!.success
-//                            )
-//                        }
+                        sharedPreferencesRepository.getUserData(response.body())
                         findNavController().navigate(R.id.action_loginCodeConfirmationFragment_to_animationFragment)
                     } else {
                         with(binding) {
@@ -196,28 +182,6 @@ class LoginCodeConfirmationFragment : Fragment(R.layout.fragment_login_code_conf
                 }
             }
         )
-    }
-
-    private fun getUserData(body: SignInConfirmResponse?) {
-        if (body != null) {
-            with(body) {
-                    val accessExpired = accessExpired
-                    val accessToken = accessToken.toString()
-                    val refreshExpired = refreshExpired
-                    val refreshToken = refreshToken.toString()
-                    val success = success
-
-                    if (accessExpired != null && refreshExpired != null) {
-                        sharedPreferencesRepository.saveUserData(
-                            accessExpired,
-                            accessToken,
-                            refreshExpired,
-                            refreshToken,
-                            success
-                        )
-                    }
-            }
-        }
     }
 }
