+++ app/src/main/java/my/doctrina/app/presentation/fragments/LoginFragment.kt
@@ -86,7 +86,7 @@ class LoginFragment : Fragment(R.layout.fragment_login) {
                         .matches()
                 ) {
 
-                    sharedPreferencesRepository.saveEmail("email", emailValidate)
+                    sharedPreferencesRepository.saveEmail(emailValidate)
 
                     userEmailVerification(emailValidate)
                 } else {
